backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openai==1.6.0
gunicorn==21.2.0
marshmallow==3.21.1
